Ensure we handle cyclic dependencies through depth
Since we're using DFS and queueing up primitive types, I can't think of
a nice way to solve depth of structs and identify if they're indeed
recursive.

Someone fix this if they can create a recursive struct solver. If the
structs were sorted in order of dependencies topologically then, the
previous algorithm works. But creating a toposort of structs are
expensive for the failing case this works better :)

Signed-off-by: Arvind Mukund <arvimuku@quicinc.com>